codecademy mini project using the html and css
This project is a mini web development task completed on the Codecademy platform. It serves as a demonstration of HTML and CSS skills acquired during the course. The web project is designed to showcase proficiency in front-end development and covers various aspects of creating a responsive and visually appealing website.